requirements.txt
functions-framework
flask
google-cloud-storage